Fix failed recur payments to have correct amount
Problem: Failed recur payments have incorrect amount set on the contribution. To replicate, 

Create a payment of $220 with $20 as the tax amount.
When the automatic recur payment is processed and failed, the new contribution is created with Total amount = $200 (instead of $220).

There are no issues with Completed payment since it calls completetransaction API to update the status.

With Failed, "create" API action is called which incorrectly updates the amount value excluding the tax amount. I'm not sure why this line was added. I think contribution amount should always be inclusive of the tax amount? Thoughts?